Extended banking Project-ATM,MiniStatement
This banking project undertakes additional functions- Deposit, Withdrawal, Transfer of funds, present Mini-Statement of banking Transactions. The project also has a special ATM function which is to work as an actual ATM should and will verify Bank account number along with Bank ATM Pin and only then will let the user undertake banking activities. The model attempts to replicate the functionality of the actual ATM and other Banking services that we come across in our day-to-day life.