File Handling+ Console Application Project CRUD
This project is  a very basic file handling and attempt to create console application project where a model museum entry and exit gate registry is created. There is potential for a lot of improvement but with projects to come in future i mean to upgrade my knowledge and make advancements to my skill of coding.