Library Management-IN-OUT time, Book ISSUE-RETURN
This is a special project that undertakes a model of library management for a university wherein a record is maintained for all library entry and exit as well as for book issue and returns. This project also implements Datetime module which automatically records the today() time and date into the records without any manual help needed.